#4 - Update sql scripts directory structure
