+++ .gitignore
@@ -6,3 +6,5 @@ database/migration_runner/migration_runner
 .vscode/
 
 config.local.yml
+
+main
