+++ database/sql_adapter/postgres/scripts/client/CreateClientTable.sql
