+++ database/sql_adapter/postgres/scripts/client/DropClientTable.sql
