+++ database/sql_adapter/postgres/scripts/client/GetClientById.sql
