+++ database/sql_adapter/postgres/scripts/migration/CreateMigrationTable.sql
