+++ database/sql_adapter/postgres/scripts/migration/GetLatestTimestamp.sql
