+++ database/sql_adapter/postgres/scripts/user/CreateUserTable.sql
