+++ database/sql_adapter/postgres/scripts/user/DeleteUser.sql
