+++ database/sql_adapter/postgres/scripts/user/DropUserTable.sql
