+++ database/sql_adapter/postgres/scripts/user/GetUserById.sql
