+++ database/sql_adapter/postgres/scripts/user/GetUserByUsername.sql
