+++ database/sql_adapter/sql_driver.go
@@ -7,25 +7,3 @@ type SQLDriver interface {
 	// GetDriverName returns the name for the driver.
 	GetDriverName() string
 }
-
-// SQLScriptRepository is an interface for fetching sql scripts.
-type SQLScriptRepository interface {
-	CreateMigrationTableScript() string
-	SaveMigrationScript() string
-	GetMigrationByTimestampScript() string
-	GetLatestTimestampScript() string
-	DeleteMigrationByTimestampScript() string
-
-	CreateUserTableScript() string
-	DropUserTableScript() string
-	SaveUserScript() string
-	GetUserByIdScript() string
-	GetUserByUsernameScript() string
-	UpdateUserScript() string
-	DeleteUserScript() string
-
-	CreateClientTableScript() string
-	DropClientTableScript() string
-	SaveClientScript() string
-	GetClientByIdScript() string
-}
